+++ kompress-core/src/commonMain/kotlin/dev/karmakrafts/kompress/DeflatingSource.kt
@@ -53,7 +53,7 @@ private class DeflatingSource( // @formatter:off
             val remaining = (byteCount - totalWritten).toInt()
             val outLimit = if (remaining < bufferSize) remaining else bufferSize
             val outBuf = if (outLimit == bufferSize) chunkBuffer else ByteArray(outLimit)
-            val written = deflater.deflate(outBuf)
+            val written = deflater.compress(outBuf)
             if (written > 0) {
                 sink.write(outBuf, 0, written)
                 totalWritten += written