+++ kompress-core/src/webMain/kotlin/dev/karmakrafts/kompress/Inflater.kt
@@ -55,7 +55,7 @@ private class InflaterImpl(raw: Boolean) : Inflater {
         get() = finalSeen && outQueue.isEmpty()
 
     @OptIn(ExperimentalUnsignedTypes::class)
-    override fun inflate(output: ByteArray): Int {
+    override fun decompress(output: ByteArray): Int {
         if (output.isEmpty()) return 0
         if (inputPending && !finalSeen) {
             val dataToPush = if (input.isNotEmpty()) input.asUByteArray().toUint8Array() else emptyUint8Array