+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 ## [Unreleased]
 
+### Added
+
+- Support for GZIP via `GZipCompressor`/`GZipDecompressor`
+- Support for ZIP via `ZipCompressor/ZipDecompressor`
+- Added `Compressor` interface
+- Added `Decompressor` interface
+
+### Changed
+
+- Updated to Kotlin Wrappers 2026.5.3
+- Updated to Android Gradle 9.2.1
+- Deprecated `Deflater.deflate` bulk compression function in favor of `Deflater.compress`
+- Deprecated `Inflater.inflate` bulk decompression function in favor of `Inflater.decompress`
+
 ## [1.4.3]
 
 ## [1.4.2]